[Loom] Remove redundant Low lower file prefixes
The low/lower package already supplies the lowering context for each of
its components. Repeating lower_ in context, report, policy, and rule
filenames made includes and build targets longer without distinguishing
ownership.

Rename the component files and tests to their package-local nouns,
update guards and includes, and teach generated rule headers to include
the renamed schema. Public loom_low_lower_* API symbols remain
unchanged.
