+++ loom/src/loom/codegen/low/lower/BUILD.bazel
@@ -29,35 +29,35 @@ _LOWER_VISIBILITY = [
 loom_cc_library(
     name = "lower",
     srcs = [
+        "context.c",
         "contract_query.c",
         "function_boundary.c",
         "lower.c",
-        "lower_context.c",
-        "lower_policy.c",
-        "lower_report.c",
-        "lower_rule_descriptor.c",
-        "lower_rule_emit.c",
-        "lower_rule_match.c",
-        "lower_rule_source_memory_match.c",
-        "lower_rule_source_memory_materialize.c",
-        "lower_rule_value.c",
         "module_state.c",
+        "policy.c",
+        "report.c",
+        "rule_descriptor.c",
+        "rule_emit.c",
+        "rule_match.c",
+        "rule_source_memory_match.c",
+        "rule_source_memory_materialize.c",
+        "rule_value.c",
         "source_plan.c",
         "source_query.c",
     ],
     hdrs = [
+        "context.h",
         "contract_query.h",
         "function_boundary.h",
         "lower.h",
-        "lower_context.h",
-        "lower_report.h",
-        "lower_rule_descriptor.h",
-        "lower_rule_emit.h",
-        "lower_rule_match.h",
-        "lower_rule_source_memory.h",
-        "lower_rule_value.h",
-        "lower_rules.h",
         "module_state.h",
+        "report.h",
+        "rule_descriptor.h",
+        "rule_emit.h",
+        "rule_match.h",
+        "rule_source_memory.h",
+        "rule_value.h",
+        "rules.h",
         "source_plan.h",
         "source_query.h",
     ],
@@ -146,8 +146,8 @@ loom_cc_test(
 )
 
 loom_cc_test(
-    name = "lower_context_test",
-    srcs = ["lower_context_test.cc"],
+    name = "context_test",
+    srcs = ["context_test.cc"],
     deps = [
         ":lower",
         "//runtime/src/iree/testing:gtest",
@@ -156,8 +156,8 @@ loom_cc_test(
 )
 
 loom_cc_test(
-    name = "lower_rule_value_test",
-    srcs = ["lower_rule_value_test.cc"],
+    name = "rule_value_test",
+    srcs = ["rule_value_test.cc"],
     deps = [
         ":lower",
         "//loom/src/loom/codegen/low/testing:source_workload",
@@ -173,8 +173,8 @@ loom_cc_test(
 )
 
 loom_cc_test(
-    name = "lower_rule_match_test",
-    srcs = ["lower_rule_match_test.cc"],
+    name = "rule_match_test",
+    srcs = ["rule_match_test.cc"],
     deps = [
         ":lower",
         "//loom/src/loom/codegen/low/testing:source_workload",
@@ -187,8 +187,8 @@ loom_cc_test(
 )
 
 loom_cc_test(
-    name = "lower_rule_emit_test",
-    srcs = ["lower_rule_emit_test.cc"],
+    name = "rule_emit_test",
+    srcs = ["rule_emit_test.cc"],
     deps = [
         ":lower",
         "//loom/src/loom/codegen/low/testing:source_workload",
@@ -265,8 +265,8 @@ loom_cc_test(
 )
 
 loom_cc_test(
-    name = "lower_report_test",
-    srcs = ["lower_report_test.cc"],
+    name = "report_test",
+    srcs = ["report_test.cc"],
     deps = [
         ":lower",
         "//loom/src/loom/codegen/low/testing:source_workload",
@@ -305,8 +305,8 @@ loom_cc_test(
 )
 
 loom_cc_test(
-    name = "lower_policy_registry_test",
-    srcs = ["lower_policy_registry_test.cc"],
+    name = "policy_registry_test",
+    srcs = ["policy_registry_test.cc"],
     deps = [
         ":lower",
         "//loom/src/loom/ops/scalar",
