+++ loom/src/loom/codegen/low/lower/lower.c
@@ -10,12 +10,12 @@
 #include "loom/analysis/contract_vector.h"
 #include "loom/analysis/kernel_async_legality.h"
 #include "loom/analysis/vector_memory_footprint.h"
+#include "loom/codegen/low/lower/context.h"
 #include "loom/codegen/low/lower/contract_query.h"
 #include "loom/codegen/low/lower/function_boundary.h"
-#include "loom/codegen/low/lower/lower_context.h"
-#include "loom/codegen/low/lower/lower_report.h"
-#include "loom/codegen/low/lower/lower_rule_emit.h"
-#include "loom/codegen/low/lower/lower_rule_source_memory.h"
+#include "loom/codegen/low/lower/report.h"
+#include "loom/codegen/low/lower/rule_emit.h"
+#include "loom/codegen/low/lower/rule_source_memory.h"
 #include "loom/codegen/low/lower/source_plan.h"
 #include "loom/codegen/low/lower/source_query.h"
 #include "loom/codegen/low/source_memory_plan.h"
