+++ loom/src/loom/codegen/low/lower/lower.h
@@ -22,8 +22,8 @@
 #include "loom/analysis/native_layout.h"
 #include "loom/analysis/symbolic_expr.h"
 #include "loom/codegen/low/descriptors.h"
-#include "loom/codegen/low/lower/lower_report.h"
 #include "loom/codegen/low/lower/module_state.h"
+#include "loom/codegen/low/lower/report.h"
 #include "loom/codegen/low/memory_access.h"
 #include "loom/error/emitter.h"
 #include "loom/error/error_defs.h"
