+++ loom/src/loom/codegen/low/lower/source_query.c
@@ -8,9 +8,9 @@
 
 #include <string.h>
 
+#include "loom/codegen/low/lower/context.h"
 #include "loom/codegen/low/lower/contract_query.h"
-#include "loom/codegen/low/lower/lower_context.h"
-#include "loom/codegen/low/lower/lower_rule_match.h"
+#include "loom/codegen/low/lower/rule_match.h"
 #include "loom/ir/module.h"
 #include "loom/ops/func/ops.h"
 #include "loom/target/low_descriptor_registry.h"
